+++ Joyride.Specflow/Steps/ScreenSteps.cs
@@ -35,7 +35,7 @@ namespace Joyride.Specflow.Steps
         {
             var text = (quotedText != String.Empty) ? quotedText : curlyText;
             string actualText = null;
-            Context.MobileApp.Do<Screen>(s => actualText = s.GetElementText(elementName));
+            Context.MobileApp.Do<Screen>(s => actualText = s.GetText(elementName));
 
             if (shouldOrShouldNot == "should")
                 Assert.That(actualText != null && actualText.CompareWith(text, compareType.ToCompareType()), Is.True,
@@ -49,7 +49,7 @@ namespace Joyride.Specflow.Steps
         public void ThenIShouldSeeElementCleared(string shouldOrShouldNot, string elementName)
         {
             string actualText = null;
-            Context.MobileApp.Do<Screen>(s => actualText = s.GetElementText(elementName));
+            Context.MobileApp.Do<Screen>(s => actualText = s.GetText(elementName));
 
             if (shouldOrShouldNot == "should")
                 Assert.That(String.IsNullOrEmpty(actualText), Is.True,
@@ -65,12 +65,12 @@ namespace Joyride.Specflow.Steps
             var elementPresent = false;
             if (shouldOrShouldNot == "should")
             {
-                Context.MobileApp.Do<Screen>(s => elementPresent = s.ElementIsPresent(elementName, TimeoutSecs));
+                Context.MobileApp.Do<Screen>(s => elementPresent = s.IsPresent(elementName, TimeoutSecs));
                 Assert.IsTrue(elementPresent);
             }
             else
             {
-                Context.MobileApp.Do<Screen>(s => elementPresent = s.ElementIsPresent(elementName, NonExistenceTimeoutSecs));
+                Context.MobileApp.Do<Screen>(s => elementPresent = s.IsPresent(elementName, NonExistenceTimeoutSecs));
                 Assert.IsFalse(elementPresent); 
             }               
         }
@@ -81,12 +81,12 @@ namespace Joyride.Specflow.Steps
             var elementVisible = false;
             if (shouldOrShouldNot == "should")
             {
-                Context.MobileApp.Do<Screen>(s => elementVisible = s.ElementIsVisible(elementName, TimeoutSecs));
+                Context.MobileApp.Do<Screen>(s => elementVisible = s.IsVisible(elementName, TimeoutSecs));
                 Assert.IsTrue(elementVisible);
             }
             else
             {
-                Context.MobileApp.Do<Screen>(s => elementVisible = s.ElementIsVisible(elementName, NonExistenceTimeoutSecs));
+                Context.MobileApp.Do<Screen>(s => elementVisible = s.IsVisible(elementName, NonExistenceTimeoutSecs));
                 Assert.IsFalse(elementVisible);
             }                
         }
@@ -134,7 +134,7 @@ namespace Joyride.Specflow.Steps
             {
                 var foundElement = false;
                 var elementName = e;
-                Context.MobileApp.Do<Screen>(s => foundElement = s.ElementIsVisible(elementName, timeoutSecs));
+                Context.MobileApp.Do<Screen>(s => foundElement = s.IsVisible(elementName, timeoutSecs));
 
                 if (shouldOrShouldNot == "should")
                     Assert.IsTrue(foundElement, "Unexpected element not visible: " + elementName);
@@ -156,7 +156,7 @@ namespace Joyride.Specflow.Steps
             {
                 var foundElement = false;
                 var elementName = e;
-                Context.MobileApp.Do<Screen>(s => foundElement = s.ElementIsPresent(elementName, timeoutSecs));
+                Context.MobileApp.Do<Screen>(s => foundElement = s.IsPresent(elementName, timeoutSecs));
 
                 if (shouldOrShouldNot == "should")
                     Assert.IsTrue(foundElement, "Unexpected element not present: " + elementName);
@@ -171,7 +171,7 @@ namespace Joyride.Specflow.Steps
             var enabled = enabledOrDisabled == "enabled";
             var actualEnabled = false;
 
-            Context.MobileApp.Do<Screen>(s => actualEnabled = s.IsEnabled(elementName));
+            Context.MobileApp.Do<Screen>(s => actualEnabled = s.IsEnabled(elementName, TimeoutSecs));
             if (shouldOrShouldNot == "should")
                 Assert.IsTrue(actualEnabled == enabled);
             else