+++ Joyride/Extensions/AppiumExtensions.cs
@@ -4,6 +4,9 @@ using System.Diagnostics;
 using System.Drawing;
 using System.Drawing.Imaging;
 using System.Linq;
+using System.Reflection;
+using System.Runtime.Remoting;
+using Joyride.Support;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Appium;
 using OpenQA.Selenium.Appium.MultiTouch;
@@ -18,26 +21,26 @@ namespace Joyride.Extensions
         //cache screen size for performance
         private static Size? _screenSize;
 
-        public static Size ScreenSize(this AppiumDriver driver)
+        public static Size ScreenSize<T>(this AppiumDriver<T> driver) where T: IWebElement
         {
             var size = _screenSize ?? driver.Manage().Window.Size;
             _screenSize = size;
             return (Size) _screenSize;
         }
 
-        public static Point ScreenCenterPoint(this AppiumDriver driver)
+        public static Point ScreenCenterPoint<T>(this AppiumDriver<T> driver) where T: IWebElement
         {
             var centerX = driver.ScreenSize().Width/2;
             var centerY = driver.ScreenSize().Height/2;
             return new Point(centerX, centerY);
         }
 
-        public static bool IsNative(this AppiumDriver driver)
+        public static bool IsNative<T>(this AppiumDriver<T> driver) where T: IWebElement
         {
             return driver.Context == NativeAppContext;
         }
 
-        public static void SwitchTo(this AppiumDriver driver, string context, int maxRetries)
+        public static void SwitchTo<T>(this AppiumDriver<T> driver, string context, int maxRetries) where T: IWebElement
         {
             var retries = 0;
             do
@@ -55,14 +58,14 @@ namespace Joyride.Extensions
                 }
             } while (retries < maxRetries);
         }
-       
-        public static string SwitchToNative(this AppiumDriver driver, int maxRetries = 3)
+
+        public static string SwitchToNative<T>(this AppiumDriver<T> driver, int maxRetries = 3) where T : IWebElement
         {
             driver.SwitchTo(NativeAppContext, maxRetries);
             return NativeAppContext;
         }
 
-        public static string SwitchToWebview(this AppiumDriver driver, int maxRetries = 3)
+        public static string SwitchToWebview<T>(this AppiumDriver<T> driver, int maxRetries = 3) where T: IWebElement
         {
             var contextCount = 1;
             var retries = 0;
@@ -87,7 +90,7 @@ namespace Joyride.Extensions
             return webViewContext;
         }
 
-        public static void DoActionInWebView(this AppiumDriver driver, Action action, int maxRetries=3)
+        public static void DoActionInWebView<T>(this AppiumDriver<T> driver, Action action, int maxRetries = 3) where T: IWebElement
         {
             if (driver.Context != NativeAppContext)
                 action();
@@ -117,25 +120,38 @@ namespace Joyride.Extensions
                 throw new ArgumentOutOfRangeException("Zoom only scales to 0.0 - 1.0.  Scale of " + scale + " is out of range.");
         }
 
-        public static void Tap(this AppiumDriver driver, Point location)
+        public static void Tap<T>(this AppiumDriver<T> driver, Point location) where T: IWebElement
         {
             new TouchAction(driver).Press(location.X, location.Y).Perform();
         }
 
-        public static void PreciseTap(this AppiumDriver driver, IWebElement element)
+        public static void PreciseTap<T>(this AppiumDriver<T> driver, IWebElement element) where T: IWebElement
         {
             driver.Tap(element.GetCenter());
         }
 
-        public static void DoubleTap(this AppiumDriver driver, IWebElement element)
+        public static void DoubleTap<T>(this AppiumDriver<T> driver, IWebElement element) where T: IWebElement
         {
            var center = element.GetCenter();
+            if (RemotingServices.IsTransparentProxy(element))
+            {
+                Trace.WriteLine("Detected proxy");
+//                var proxy = RemotingServices.GetRealProxy(element);
+//                var cachedElement = (IWebElement) Util.GetMemberValue(proxy, "cachedElement", BindingFlags.NonPublic);
+//                new TouchAction(driver).
+//                    Tap(cachedElement, center.X, center.Y, 2)
+//                    .Perform();            
+
+            }
+                  
+
            new TouchAction(driver).
                Tap(element, center.X, center.Y, 2)
                .Perform();            
         }
-        
-        public static void TapAndHold(this AppiumDriver driver, IWebElement element, int seconds, bool precise=false)
+
+        public static void TapAndHold<T>(this AppiumDriver<T> driver, IWebElement element, int seconds, bool precise = false)
+            where T : IWebElement
         {
             if (precise)
             {
@@ -156,8 +172,9 @@ namespace Joyride.Extensions
             }
         }
 
-        public static void Swipe(this AppiumDriver driver, Direction direction, Size dimension,
-             double scale=1.0, long durationMilliSecs = 500, int originX=0, int originY=0)
+        public static void Swipe<T>(this AppiumDriver<T> driver, Direction direction, Size dimension,
+             double scale=1.0, long durationMilliSecs = 500, int originX=0, int originY=0) 
+            where T: IWebElement
         {
             var center = new Point(originX + dimension.Width / 2, originY + dimension.Height / 2);
             var startDeltaX = dimension.Width / 3;
@@ -203,14 +220,16 @@ namespace Joyride.Extensions
                 .Release().Perform();
         }
 
-        public static void Swipe(this AppiumDriver driver, Direction direction, double scale=1.0, long durationMilliSecs=500)
+        public static void Swipe<T>(this AppiumDriver<T> driver, Direction direction, double scale = 1.0, long durationMilliSecs = 500)
+            where T: IWebElement
         {
             EnsureScaleRange(scale);
             EnsureNotInOutDirection(direction);
             Swipe(driver, direction, driver.ScreenSize(), scale, durationMilliSecs);
         }
-       
-        public static void Swipe(this AppiumDriver driver, IWebElement element, Direction direction, double scale = 1.0, long durationMilliSecs = 500)
+
+        public static void Swipe<T>(this AppiumDriver<T> driver, IWebElement element, Direction direction, double scale = 1.0, long durationMilliSecs = 500)
+            where T : IWebElement
         {
             EnsureNotInOutDirection(direction);
             EnsureScaleRange(scale);
@@ -256,7 +275,8 @@ namespace Joyride.Extensions
                 .Perform();
         }
 
-        private static Size CalculateVisibleElementSize(AppiumDriver driver, IWebElement element)
+        private static Size CalculateVisibleElementSize<T>(AppiumDriver<T> driver, IWebElement element)
+            where T : IWebElement
         {
             var size = element.Size;
             var lowerRight = new Point(element.Location.X + size.Width, element.Location.Y + size.Height);
@@ -267,7 +287,8 @@ namespace Joyride.Extensions
             return size;
         }
 
-        public static void PinchToZoom(this AppiumDriver driver, Direction direction, double scale = 1.0)
+        public static void PinchToZoom<T>(this AppiumDriver<T> driver, Direction direction, double scale = 1.0) 
+            where T: IWebElement
         {
             EnsureScaleRange(scale);          
             EnsureInOutDirection(direction);
@@ -331,7 +352,8 @@ namespace Joyride.Extensions
             return directionToSwipe;
         }
 
-        public static void Scroll(this AppiumDriver driver, Direction direction, double scale=1.0, long durationMilliSecs = 2000)
+        public static void Scroll<T>(this AppiumDriver<T> driver, Direction direction, double scale = 1.0, long durationMilliSecs = 2000)
+            where T : IWebElement
         {
             EnsureNotInOutDirection(direction);
             EnsureScaleRange(scale);            
@@ -339,7 +361,7 @@ namespace Joyride.Extensions
             driver.Swipe(directionToSwipe, scale, durationMilliSecs);
         }
 
-        public static bool ElementWithinBounds(this AppiumDriver driver, IWebElement element)
+        public static bool ElementWithinBounds<T>(this AppiumDriver<T> driver, IWebElement element) where T: IWebElement
         {
             if (element == null)
                 return false;
@@ -360,7 +382,8 @@ namespace Joyride.Extensions
             return true;
         }
 
-        public static void CaptureScreenshot(this AppiumDriver driver, string pathAndFilename, ImageFormat format)
+        public static void CaptureScreenshot<T>(this AppiumDriver<T> driver, string pathAndFilename, ImageFormat format)
+            where T : IWebElement
         {
             try
             {
@@ -374,8 +397,8 @@ namespace Joyride.Extensions
             }
         }
 
-        public static void SwipeFromEdge(this AppiumDriver driver, Direction direction, long durationMillsecs = 1000,
-            double scale = 1.0, double offset=0)
+        public static void SwipeFromEdge<T>(this AppiumDriver<T> driver, Direction direction, long durationMillsecs = 1000,
+            double scale = 1.0, double offset = 0) where T : IWebElement
         {
             EnsureNotInOutDirection(direction);
             double startX= 1.0, endX = 1.0, startY = 1.0, endY = 1.0;
@@ -418,8 +441,9 @@ namespace Joyride.Extensions
                 .Perform();
 
         }
-        
-        public static void PullScreen(this AppiumDriver driver, Direction direction, long durationMillsecs=1000, double scale=1.0, double offset=0)
+
+        public static void PullScreen<T>(this AppiumDriver<T> driver, Direction direction, long durationMillsecs = 1000, double scale = 1.0, double offset = 0)
+            where T : IWebElement
         {
 
             EnsureNotInOutDirection(direction);
@@ -428,7 +452,8 @@ namespace Joyride.Extensions
             SwipeFromEdge(driver, direction, durationMillsecs, scale, offset);
         }
 
-        public static void DragAndDrop(this AppiumDriver driver, IWebElement fromElement, IWebElement toElement, long durationMilliSecs = 1000)
+        public static void DragAndDrop<T>(this AppiumDriver<T> driver, IWebElement fromElement, IWebElement toElement, long durationMilliSecs = 1000)
+              where T : IWebElement
         {
             if (fromElement == null || toElement == null)
                 throw new NoSuchElementException("Unable to perform drag action due to missing elements");