+++ Joyride/Platforms/MobileApplication.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Diagnostics;
 using Joyride.Extensions;
+using OpenQA.Selenium;
 using OpenQA.Selenium.Appium;
 
 namespace Joyride.Platforms
@@ -13,7 +14,7 @@ namespace Joyride.Platforms
         protected int TransitionDelayMs = 0;
         abstract public string Identifier { get; }
         public Screen Screen { get { return CurrentScreen; }}
-        protected AppiumDriver Driver { get { return RemoteMobileDriver.GetInstance(); } }
+        protected AppiumDriver<IWebElement> Driver { get { return RemoteMobileDriver.GetInstance(); } }
 
         public virtual void Launch()
         {
@@ -47,7 +48,7 @@ namespace Joyride.Platforms
         {
             var anyScreenOrInterface = CastScreen<T>();
             var beforeTransition = CurrentScreen;
-            CurrentScreen = func(anyScreenOrInterface);
+            CurrentScreen = func(anyScreenOrInterface) ?? beforeTransition;
 
             if (CurrentScreen != beforeTransition)
             {